Remove discovery module
