+++ kotest-runner/kotest-runner-junit5/src/jvmMain/kotlin/io/kotest/runner/junit/platform/KotestEngineDescriptor.kt
@@ -1,10 +1,10 @@
 package io.kotest.runner.junit.platform
 
 import io.kotest.core.extensions.Extension
-import io.kotest.engine.extensions.DescriptorFilter
 import io.kotest.core.log
 import io.kotest.core.spec.Spec
 import io.kotest.engine.descriptors.toDescriptor
+import io.kotest.engine.extensions.DescriptorFilter
 import io.kotest.engine.test.names.FallbackDisplayNameFormatter
 import io.kotest.runner.junit.platform.gradle.GradleClassMethodRegexTestFilter
 import org.junit.platform.engine.UniqueId
@@ -16,21 +16,19 @@ class KotestEngineDescriptor(
    val classes: List<KClass<out Spec>>,
    val testFilters: List<DescriptorFilter>,
    val extensions: List<Extension>, // extensions can be added via junit configuration parameters
-   val error: Throwable?, // an error during discovery
 ) : EngineDescriptor(id, KotestJunitPlatformTestEngine.ENGINE_NAME) {
    // Only reports dynamic children (see ExtensionExceptionExtractor) if there are no test classes to run.
    override fun mayRegisterTests(): Boolean = classes.isEmpty()
 }
 
 internal fun createEmptyEngineDescriptor(id: UniqueId): KotestEngineDescriptor {
-   return KotestEngineDescriptor(id, emptyList(), emptyList(), emptyList(), null)
+   return KotestEngineDescriptor(id, emptyList(), emptyList(), emptyList())
 }
 
 internal fun createEngineDescriptor(
    uniqueId: UniqueId,
    specs: List<KClass<out Spec>>,
    gradleClassMethodTestFilter: GradleClassMethodRegexTestFilter?,
-   error: Throwable?,
    extensions: List<Extension>, // extensions can be added via junit configuration parameters
 ): KotestEngineDescriptor {
 
@@ -39,7 +37,6 @@ internal fun createEngineDescriptor(
       classes = specs,
       testFilters = listOfNotNull(gradleClassMethodTestFilter),
       extensions = extensions,
-      error = error,
    )
 
    val formatter = FallbackDisplayNameFormatter.default()
